refactor : controller 자동차 이름 입력 관련 리팩토링
